clarify go mod init
We indicate that a go module should be initialised.
This should be accompanied by a brief explanation why.
This command plays an important role in helping manage
project dependencies.
While it is not critical for a hello world program it is
generally good practice and essential for any non-trivial
project.