add dedicated section on modifying PATH
Mofifying the path has to be mentioned multiple times in the guide.
It is a good idea make a generic comment about this step up-front.
This increases the odds that readers will note the point and it
means we do not have to belabor the point elsewhere for fear that
someone skimming a specific section will miss it.