Update toolchain availability reference (#42)
* chore: remove unused shared links document

The shared links document is unused, because it requires
reStructuredText to use. We use myST which does not have a
similar feature.

The document was just part of the Canonical Sphinx docs starter pack.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* feat: seperate availability ref in seperate docs

Seperate the reference of available versions for the toolchains into
seperate documents by toolchain.

This commit also applies minor updates to the availability matracies.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* feat: add .NET backports/previews PPA reference

Add a smalll reference outlining the .NET backports and previews PPA.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* fix: spelling warnings

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* chore: remove availability row from index table

Remove the row with the links to the toolchain availability pages from
the index page of the docs.

This is based on the feedback of our technical author Robert Krátký:

> I think extending the table this way is not necessary -- let's keep
> the landing page uncluttered. While the reference info is useful, it's
> not the main content of the docs set. I'd like the table to link
> (only) to the two basic topics:
> - how to get the toolchains
> - how to start using them

Co-authored-by: Robert Krátký <robert.kratky@canonical.com>
Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* feat: move .NET snap install args reference

Move the reference for valid components and versions to install to the
.NET availability reference and link to it from the how-to/dotnet-setup
article.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* refactor: replace <span> elements with sphinx role

Replace the <span> elements to use a custom css class by defining a
spinx role and using this instead.

This makes the source text much more readable and maintainable.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* refactor: replace default-package role with new table row

Instead of using a custom default-package role to highlight the
llvm-default version in orange, we replace it by using a sperate row
that shows the default version.

This was suggested by Robert Krátký.

Co-authored-by: Robert Krátký <robert.kratky@canonical.com>
Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

* chore: revert accidentially removed character

A previous commit accidentially removed a `|` from index.md at the
end of a table column.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>

---------

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>
Co-authored-by: Robert Krátký <robert.kratky@canonical.com>