chore: remove unused dependency sphinx-prompt
The sphinx-prompt dependency was introduced with PR #22. PR #36 replaced
the used feature with a similar feature included in the canonical_sphinx
extension. Therefore the depenency is no longer used and can be removed.

Furthermore the sphinx-prompt dependency recently started causes an
issue when trying to build the documentation:

> Extension error:
> Could not import extension sphinx-prompt (exception: No module named 'sphinx-prompt')
> Sphinx exited with exit code: 2

Apparently the extension name was renamed from "sphinx-prompt" to
"sphinx_prompt".

We could easily fix the issue by appying the name change to
docs/conf.py, but removing the dependency makes more sense
in this context.

PR #22:  https://github.com/canonical/ubuntu-for-developers-docs/pull/22
PR #36: https://github.com/canonical/ubuntu-for-developers-docs/pull/36

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>