style: use dropdown box for install software-properties-common tip
Use a dropdown box for the tip to apt install the package
software-properties-common if the command add-apt-repository(1)
cannot be found.

This tip is not important for the majority of all readers, but takes up
a non insignificant amount of space. IMHO, putting this tip into a
dropdown box makes it less visually noisy.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>